osd_types: use uint64_t for ObjectExtent offsets and lengths
This is just client in-memory state, and allows us to address objects >4GiB,
to match the existing librados/Objecter interfaces.

+++ src/client/SyntheticClient.cc
@@ -1677,7 +1677,7 @@ int SyntheticClient::dump_placement(string& fn) {
     int osd = client->osdmap->get_pg_primary(client->osdmap->object_locator_to_pg(i->oid, i->oloc));
 
     // run through all the buffer extents
-    for (map<__u32,__u32>::iterator j = i ->buffer_extents.begin();
+    for (map<uint64_t, uint64_t>::iterator j = i ->buffer_extents.begin();
 	 j != i->buffer_extents.end(); ++j) {
       
       dout(0) << "OSD " << osd << ", offset " << (*j).first <<
